Move game settings to common.
