+++ front/src/common/game_settings.scss
